topology: extract TopologyGraphSvg; reuse on dashboard
TopologyGraphSvg is a pure prop-based SVG renderer — graph required,
everything else optional with safe defaults. TopologyGraph becomes a
thin context wrapper. Dashboard drops its 80-line inline SVG copy and
uses <TopologyGraphSvg graph={graph} /> instead.

This fixes two dashboard issues for free: edge de-duplication (via the
shared buildTopoGraph) and stale node layout (circle/text positions
were written for old NODE_H=38, now consistent at 50).
